pyproject.toml
[project]
name = "stellarmesh"
version = "0.4.2"
authors = [{ name = "Alex Koen" }]
description = "Gmsh wrapper and DAGMC geometry creator."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gmsh",
    "numpy",
    "pymoab",
    "pymmg",
    "cadquery-ocp~=7.7.2.1",
    "meshio[all]",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"
extend-include = ["*.ipynb"]


[tool.ruff.lint]
select = [
    "E",
    "W",
    "I",
    "N",
    "A",
    "B",
    "F",
    "D",
    "FBT",
    "C4",
    "SIM",
    "TD",
    "PL",
    "TRY",
    "NPY",
    "PERF",
    "RUF",
]
ignore = [
    "TRY003",  # Exception with string (vs. defining custom exception)
    "PLR2004", # Magic number
]
pydocstyle.convention = "google"


[tool.pyright]
pythonVersion = "3.10"
